move sync to import
