+++ tests/store/api/test_store_import_func.py
@@ -2,9 +2,9 @@
 
 from cg.store import Store
 from cg.store.api.import_func import prices_are_same, versions_are_same, \
-    import_application_versions, get_raw_data_from_xl, get_tag_from_raw_version, \
-    add_version_from_raw, get_workbook_from_xl, import_applications, get_datemode_from_xl, \
-    get_tag_from_raw_application
+    import_application_versions, _get_raw_data_from_xl, _get_tag_from_raw_version, \
+    add_version_from_raw, _get_workbook_from_xl, import_applications, _get_datemode_from_xl, \
+    _get_tag_from_raw_application, import_apptags
 
 
 def test_prices_are_same_int_and_int():
@@ -63,14 +63,14 @@ def test_versions_are_same(store: Store, application_versions_file):
     # GIVEN an excel price row
     # same price row committed to the database
     add_applications(store, application_versions_file)
-    raw_version = get_raw_data_from_xl(application_versions_file)[0]
-    tag = get_tag_from_raw_version(raw_version)
+    raw_version = _get_raw_data_from_xl(application_versions_file)[0]
+    tag = _get_tag_from_raw_version(raw_version)
     application_obj = store.application(tag)
     sign = 'DummySign'
-    workbook = get_workbook_from_xl(application_versions_file)
+    workbook = _get_workbook_from_xl(application_versions_file)
     db_version = add_version_from_raw(application_obj, None, raw_version,
                                       sign, store, workbook)
-    datemode = get_datemode_from_xl(application_versions_file)
+    datemode = _get_datemode_from_xl(application_versions_file)
 
     # WHEN calling versions are same
     should_be_same = versions_are_same(db_version, raw_version, datemode)
@@ -83,15 +83,15 @@ def test_versions_are_not_same(store, application_versions_file):
     # GIVEN an excel price row
     # NOT same price row committed to the database
     add_applications(store, application_versions_file)
-    raw_version = get_raw_data_from_xl(application_versions_file)[0]
-    tag = get_tag_from_raw_version(raw_version)
+    raw_version = _get_raw_data_from_xl(application_versions_file)[0]
+    tag = _get_tag_from_raw_version(raw_version)
     application_obj = store.application(tag)
     sign = 'DummySign'
-    workbook = get_workbook_from_xl(application_versions_file)
+    workbook = _get_workbook_from_xl(application_versions_file)
     db_version = add_version_from_raw(application_obj, None, raw_version,
                                       sign, store, workbook)
-    datemode = get_datemode_from_xl(application_versions_file)
-    another_raw_version = get_raw_data_from_xl(application_versions_file)[1]
+    datemode = _get_datemode_from_xl(application_versions_file)
+    another_raw_version = _get_raw_data_from_xl(application_versions_file)[1]
 
     # WHEN calling versions are same
     should_not_be_same = versions_are_same(db_version, another_raw_version, datemode)
@@ -124,6 +124,105 @@ def test_application(applications_file, store: Store):
     assert all_applications_exists(store, applications_file)
 
 
+def test_sync_microbial_orderform_dryrun(base_store: Store, microbial_orderform):
+    # GIVEN a microbial orderform and a store where all the apptags exists half some inactive and
+    # some active
+    ensure_applications(base_store, ['MWRNXTR003', 'MWGNXTR003', 'MWMNXTR003', 'MWLNXTR003'],
+                        ['MWXNXTR003', 'VWGNXTR001', 'VWLNXTR001'])
+    prep_category = 'mic'
+    sign = 'PG'
+    activate = False
+    inactivate = False
+    active_mic_apps_from_start = base_store.applications(category=prep_category,
+                                                         archived=False).count()
+    inactive_mic_apps_from_start = base_store.applications(category=prep_category,
+                                                           archived=True).count()
+
+    # WHEN syncing app-tags in that orderform
+    import_apptags(store=base_store, excel_path=microbial_orderform,
+                   prep_category=prep_category, sign=sign, activate=activate,
+                   inactivate=inactivate, sheet_name='Drop down list', tag_column=2)
+
+    # THEN same number of active and inactive mic applications in status database
+    active_mic_apps_after_when = base_store.applications(category=prep_category,
+                                                         archived=False).count()
+    inactive_mic_apps_after_when = base_store.applications(category=prep_category,
+                                                           archived=True).count()
+    assert active_mic_apps_from_start == active_mic_apps_after_when
+    assert inactive_mic_apps_from_start == inactive_mic_apps_after_when
+
+
+def test_sync_microbial_orderform_activate(base_store: Store, microbial_orderform):
+    # GIVEN a microbial orderform and a store where all the apptags exists half some inactive and
+    # some active
+    ensure_applications(base_store, ['MWRNXTR003', 'MWGNXTR003', 'MWMNXTR003', 'MWLNXTR003'],
+                        ['MWXNXTR003', 'VWGNXTR001', 'VWLNXTR001'])
+    prep_category = 'mic'
+    sign = 'PG'
+    activate = True
+    inactivate = False
+    active_mic_apps_from_start = base_store.applications(category=prep_category,
+                                                         archived=False).count()
+    inactive_mic_apps_from_start = base_store.applications(category=prep_category,
+                                                           archived=True).count()
+
+    # WHEN syncing app-tags in that orderform
+    import_apptags(store=base_store, excel_path=microbial_orderform,
+                   prep_category=prep_category, sign=sign, activate=activate,
+                   inactivate=inactivate, sheet_name='Drop down list', tag_column=2)
+
+    # THEN more active mic applications in status database and same inactive
+    active_mic_apps_after_when = base_store.applications(category=prep_category,
+                                                         archived=False).count()
+    inactive_mic_apps_after_when = base_store.applications(category=prep_category,
+                                                           archived=True).count()
+    assert active_mic_apps_from_start < active_mic_apps_after_when
+    assert inactive_mic_apps_from_start > inactive_mic_apps_after_when
+
+
+def test_sync_microbial_orderform_inactivate(base_store: Store, microbial_orderform):
+    # GIVEN a microbial orderform and a store where all the apptags exists half some inactive and
+    # some active
+    ensure_applications(base_store, ['MWRNXTR003', 'MWGNXTR003', 'MWMNXTR003', 'MWLNXTR003',
+                                     'dummy_tag'],
+                        ['MWXNXTR003', 'VWGNXTR001', 'VWLNXTR001'])
+    prep_category = 'mic'
+    sign = 'PG'
+    activate = False
+    inactivate = True
+    active_mic_apps_from_start = base_store.applications(category=prep_category,
+                                                         archived=False).count()
+    inactive_mic_apps_from_start = base_store.applications(category=prep_category,
+                                                           archived=True).count()
+
+    # WHEN syncing app-tags in that orderform
+    import_apptags(store=base_store, excel_path=microbial_orderform,
+                   prep_category=prep_category, sign=sign, activate=activate,
+                   inactivate=inactivate, sheet_name='Drop down list', tag_column=2)
+
+    # THEN same number of active and more inactive mic applications in status database
+    active_mic_apps_after_when = base_store.applications(category=prep_category,
+                                                         archived=False).count()
+    inactive_mic_apps_after_when = base_store.applications(category=prep_category,
+                                                           archived=True).count()
+    assert active_mic_apps_from_start > active_mic_apps_after_when
+    assert inactive_mic_apps_from_start < inactive_mic_apps_after_when
+
+
+def ensure_applications(base_store: Store, active_applications: list, inactive_applications: list):
+    """Create some requested applications for the tests """
+    for active_application in active_applications:
+        if not base_store.application(active_application):
+            base_store.add_commit(base_store.add_application(active_application, 'mic',
+                                                             'dummy_description',
+                                                             is_archived=False))
+
+    for inactive_application in inactive_applications:
+        if not base_store.application(inactive_application):
+            base_store.add_commit(base_store.add_application(inactive_application, 'mic',
+                                                             'dummy_description', is_archived=True))
+
+
 def ensure_application(store, tag):
     """Ensure that the specified application exists in the store"""
     application = store.application(tag=tag)
@@ -139,22 +238,22 @@ def ensure_application(store, tag):
 def add_applications(store, application_versions_file):
     """Ensure all applications in the xl exists"""
 
-    raw_versions = get_raw_data_from_xl(application_versions_file)
+    raw_versions = _get_raw_data_from_xl(application_versions_file)
 
     for raw_version in raw_versions:
-        tag = get_tag_from_raw_version(raw_version)
+        tag = _get_tag_from_raw_version(raw_version)
         ensure_application(store, tag)
 
 
 def get_prices_from_store(store, raw_price):
     """Gets all versions for the specified application"""
-    tag = get_tag_from_raw_version(raw_price)
+    tag = _get_tag_from_raw_version(raw_price)
     return store.application(tag).versions
 
 
 def get_application_from_store(store, raw_application):
     """Gets the specified application"""
-    tag = get_tag_from_raw_application(raw_application)
+    tag = _get_tag_from_raw_application(raw_application)
     return store.application(tag)
 
 
@@ -172,8 +271,8 @@ def exists_version_in_store(raw_price, store, datemode):
 
 def all_versions_exists_in_store(store, excel_path):
     """Check if all versions in the excel exists in the store"""
-    raw_versions = get_raw_data_from_xl(excel_path)
-    datemode = get_datemode_from_xl(excel_path)
+    raw_versions = _get_raw_data_from_xl(excel_path)
+    datemode = _get_datemode_from_xl(excel_path)
     for raw_version in raw_versions:
         if not exists_version_in_store(raw_version, store, datemode):
             return False
@@ -183,7 +282,7 @@ def all_versions_exists_in_store(store, excel_path):
 
 def all_applications_exists(store, applications_file):
     """Check if all applications in the excel exists in the store"""
-    raw_applications = get_raw_data_from_xl(applications_file)
+    raw_applications = _get_raw_data_from_xl(applications_file)
 
     for raw_application in raw_applications:
         if not exists_application_in_store(raw_application, store):
